selectors: add logging to getPrivateMessages
Sentry reports indicate that `getPrivateMessages` is sometimes
returning a list containing `undefined`. Prevent these from being
returned, and log IDs that would have caused them to Sentry.

This ameliorates the known symptoms of issue #3749, but does not
actually fix it.

It would be convenient, and perhaps not entirely wrong, to lay blame
for this error on the unsoundness of indexer-element types [1]... but
it actually should be the case that all indices in `narrows` are also
found in `messages`.

[1] Specifically, the fact that

    (a: {[key: string]: T})["string"]

is inferred to be of type `T`, but may actually be `undefined`. (This
unsoundness is, sadly, shared by TypeScript and Flow.)