realm screen [nfc]: Refactor styles for editability.
We're going to be making changes to the styles in the course of
making changes to the UI of this screen.  Start by pulling the
details of the styles into this file itself, with names specific to
where we use them, so we can simply edit them freely.

One pattern we've used elsewhere for doing that is to have `styles`
as an instance property defined just above `render`.  But it turns
out that that doesn't interact properly with hot reload, which is an
essential development feature for working on a UI.  So instead we
define `styles` inside `render`, just above the `return` statement
with the big blob of JSX.

I think historically we've often avoided writing styles in a pattern
like this because it means the `style` props we pass will have new,
non-identical values each time our `render` is called.  In
principle, that could defeat some React optimizations at our child
components that can avoid further `render` calls at the children.
But this looks like premature optimization to me: we've never had
evidence that this pattern makes a real difference to the app's
performance, and for a component like this (with a fixed handful of
children, no giant lists) it seems pretty unlikely that it would.
So I'm not inclined to continue contorting the code to cater to it.