StreamScreen: Default stream notification setting to user setting.
Make subscription.push_notifications null | boolean, following the
migration in https://github.com/zulip/zulip/pull/11507. If the user
has not specified true or false for a stream's notifications, then
push_notifications will be null, and the Notifications switch will
fall back to the streamNotification user setting. Signal to the
server that we will accept null for subscription.push_notifications
by passing the following with the /register request:

client_capabilities: {
    notification_settings_null: true,
}

There is currently no way to go back to the null setting on a stream
if you have already indicated true or false.

Fixes: #3449.