Add migration to strip trailing slashes on realm in all existing accounts.
Fixes: #3567.

A trailing slash on a realm URL interferes with /events calls.

The parent commit handles multiple trailing slashes only on fresh
URL inputs. A migration is needed for users whose cached realm URLs
have multiple.

Greg points out that we will want to take input from the server on
what the canonical realm URL is, and use that.