ios build: Remove CocoaPods reference from Xcode workspace.
Some time ago, there was an attempt to convert to CocoaPods. [0] This
was mostly removed a few months later, [1] but a few dormant fragments
remained.

However, we're about to switch over to CocoaPods (again?) as part of
the React Native 0.60 upgrade -- and one of those old fragments keeps
being awakened when switching between the RN 0.60 experimental branch
and our current one, causing frustrating build failures.

Sweep it away.

[0] added:   aded466bcdac5783e33bb61c627af0d9472a2c13, 2017-03-19
[1] removed: 4096e712dd63a909825e9d707e2b6b546abdbf98, 2017-07-18