message actions [nfc]: Clarify which narrows offer "edit message".
This conditional does not mean the same thing as the old version!
The old version produced true for search narrows.

That would have been a bug, but... when we show a search narrow,
we actually pass `narrow={HOME_NARROW}` to `MessageList` (in
`SearchMessagesCard.js`), and so the `narrow` value that makes it
here is the home narrow.  So the bug is only latent, and this
commit makes no functional change.  Still good to fix latent bugs,
and to be explicit about the narrows where we *are* prepared to
offer the edit-message UI rather than try to list where we aren't.