unreads [nfc]: Unpack narrow structurally in getUnreadCountForNarrow.
Use `caseNarrow` to let the specifics of narrow "operands", etc.,
stay an implementation detail of the `Narrow` type.

This also gives us type-checking that we've thought of all the types
of narrows here... and exposes that we had not in fact thought of the
all-PMs narrow!  That's purely a latent bug: despite its general name,
this selector is used only to feed one part of the UI, namely the
"N unreads" banner above a message list, and we don't expose the
all-PMs narrow as a message list you can actually view.  But it's good
to make explicit so we aren't surprised in the future.

The diff is best read with `git diff -b`, as it causes some re-indents.