narrow [nfc]: Rename isPmNarrow, from isPrivateOrGroupNarrow.
As discussed in the parent commit, this category is typically the
more natural one than being specific about 1:1 vs. group PMs.
So, give this one the simpler, shorter name.

The new names also make the relationship between these predicates
clear in a way it wasn't when the term "private" mean 1:1 PMs:

  isPmNarrow
  is1to1PmNarrow
  isGroupPmNarrow

Made the change automatically, like so:

  $ perl -i -0pe '
        s/isPrivateOrGroupNarrow/isPmNarrow/g
      ' src/**/*.js

  $ tools/fmt