narrow [nfc]: Rename a few more "group narrow" -> "group PM narrow".
This goes along with our renaming a few commits ago of isGroupNarrow
to isGroupPmNarrow.  Here again, it just makes it a bit more explicit
that these are a specific case of the general class of PM narrows.

(Probably all these functions shouldn't exist, either -- the users, or
perhaps user IDs, should just get passed down as props to the couple
of components these are here for.  But that's a different refactor.)