msglist [nfc]: Use `null`, not 0, as "no such first-unread" sentinel.
Zero is never a great choice for a "no such item" sentinel where
a `number` is expected -- it makes it likely the sentinel will be
implicitly treated like just another number.

Here, it turns out that where the value ultimately gets consumed
(in `scrollToAnchor`), it has the effect of causing no message
node to be found... so that we scroll to the *bottom*.  Make that
path through the control flow explicit.

In this case, 0 is an especially confusing value to use for "there
is no unread message" because this value is called "anchor"... and
in other places called "anchor", we use 0 (as FIRST_UNREAD_ANCHOR)
to mean the first unread message!