outbox [nfc]: Send based on `display_recipient`, not `narrow`.
This was the last place we were using the `narrow` property on
outbox messages, so taking it out will mean we can remove that
property in the next commit.  Which is good because (a) the `Narrow`
type is a great big mess, and the fewer different parts of the code
are using it, the fewer need to be migrated to a cleaned-up version
of it; and (b) it was never an appropriate type to use to describe a
single message's destination.

I don't love the "display recipient" values either, but their
structure is at least less wild than Narrow, and their semantic
range is an appropriate match for "where this individual message
is sent".  And whenever we decide it's that type's turn to get
straightened out, this usage should fit in nicely to be swept up
with all the others, because it comes right after a check of the
display_recipient's friend the `type: 'stream' | 'private'` field.

This commit makes a small change in the data we send to the server
for sending a PM:
 * The old code would leave out the self user, except on self-PMs.
 * This new code will always include the self user.

Both conventions are found in different areas of the server and
webapp.  The new one here matches the display_recipient the server
sends on PMs, and is simpler.  In any event the server doesn't
care, and that's a deliberate API choice: we want to be sure to
make easy things easy when writing bots, and allowing a range of
natural ways to say who to send a message to is a key part of that.

Concretely in the server implementation, since earliest times it
drops dupes and does its own normalization whether to include the
sender, so it's essentially always lived up to that plan.  The
relevant logic lives today (as of zulip/zulip@620e9cbf7) in the
function get_recipient_from_user_profiles; and though it's moved
around and been refactored, its logic is essentially unchanged since
zulip/zulip@f851d9437, in 2012 a few months after Zulip began.