unread cards: Remove redundant loading indicator on session.loading.
The loading indicator on an empty unread messages list during
`session.loading` is redundant, following the introduction of
LoadingBanner in this series of commits. Now, we can display the "No
unread messages" text as our best guess at the current state, even
though we know it's stale, since we're also displaying a loading
indicator above.