selectors [nfc]: Cut linear-searching getSubscriptionForId.
This fixes a case of #3339.  Additionally, this is an instance
of the bug-causing NULL_* pattern, described in nullObjects.js.
We can limit the damage by pushing it down to code that's less
general and is one step closer to the particular UI code where
the decision how to handle a missing item belongs.