no-messages [nfc]: Distinguish different "no messages" names.
Confusingly, we had a `noMessages` boolean and a `NoMessages`
component -- and `noMessages` *almost* aligned with "do we
show `NoMessages`", but not quite.

Try to clear that up by, first, expanding the boolean's name
to convey more of its meaning, which is "we *have* no messages";
and, second, giving a distinct name `sayNoMessages` to the
condition "we're going to say 'No messages' to the user".