notifications: Don't register for Android notifications on iOS.
This fixes a bad refactor in commit 8ee1793c1, back in 2019-01,
which dropped an `else` when it factored this logic out as a method.

The consequence of the bug is that on iOS, we've been listening both
for the notification-related events we actually want, and for those
of the form that we would want to listen for on Android.

This would be harmless if the extra subscription never saw any
events.  But, in fact, the notification-related events we get on
Android are sent through `DeviceEventEmitter` by our Android-native
notifications code; and on iOS, the Wix `react-native-notifications`
library we rely on also uses `DeviceEventEmitter`, but *internally*.
It expects you to hook up to its own event source, which it
implements as a thin layer that does some unwrapping of the events
its iOS-native code sends over `DeviceEventEmitter`.

And the names we use over the `DeviceEventEmitter` channel on
Android are chosen to match the names the Wix library offers on its
own event source... which in turn match the names it uses, for its
internal events, on `DeviceEventEmitter`.  As a result, because we
listen for both types of events, our get-device-token callback has
been receiving both hex strings and wrapper-objects containing them
(of the form `{ deviceToken: "bbd07c4fc..."  }`).

Needless to say, this has not played well with our notifications code.
The result has historically been that we've sent "[Object object]" as
our device token upon receiving the wrapper-object; while this is
*usually* clobbered immediately thereafter by the correct value, the
vagaries of asynchrony mean that it can't be relied on to be.

The fix is simple: put back the `else` dropped in 8ee1793c1, so that
we only listen for the events we're meant to see.

[greg: revised commit message]

GitHub-PR: https://github.com/zulip/zulip-mobile/pull/3910
Fixes: #3672