messageActionSheet tests: Clean up some stream names + IDs.
The Subscription object has both a stream name and a stream ID, and
so to keep the test data well-formed these should be kept aligned in
referring to the same stream.  Similarly the Message object.

Right now to make the test pass it's enough to set the stream name,
because we don't use the stream ID; but that's a case of #3918 and
should change, after which the mismatch will be a live cause of test
failures.

We could do this by passing a stream object to `eg.makeSubscription`
and to `eg.streamMessage`.  But the most natural one to use is
`eg.stream`; and the example data already refers to `eg.stream` by
default in all the natural places, among them `eg.streamMessage()`
and `eg.subscription`, so simply use that.