ios notif: Consume "background queue" so notifs navigate from background.
This is a short-term fix that relies on a deprecated API. We should
upgrade to a newer version of react-native-notifications or use
`@react-native-community/push-notification-ios` instead.

For docs on this as of the library version we're using (1.5.0 on NPM,
which has no Git tag but corresponds to commit 882775fb5), see:
  https://github.com/wix/react-native-notifications/blob/882775fb5/docs/notificationsEvents.md

The API was removed in release 2.0.6-shapshot-8 with commit:
  https://github.com/wix/react-native-notifications/commit/800669120
presumably in response to:
  https://github.com/wix/react-native-notifications/issues/339

Process an opened notification from the "background queue" so our
handleNotificationOpen will run, and the app will navigate, for a
notification opened while the app was in the background.

Fixes: #3647