redux types: Make zulipVersion string | null.
Keys with value `undefined` don't survive round-trips into JSON, and
the zulipVersion will be persisted in JSON. Use `null`, to maintain
that Account objects always have the same number of keys, and we
don't have to worry about the distinction between an optional
`string` property and a property that's `string | void`.

The same strategy is in use for the ackedPushToken.

As a side-effect, it's much simpler to distinguish between the
second and third behaviors of `makeAccount` in the list below, for
tests. We don't need the `shouldHaveZulipVersion` param anymore:

1. Supply our own valid string value for the `zulipVersion` field in
the returned Account object

2. Make the returned Account object have an empty `zulipVersion`

3. Make the returned Account object have a default, valid string
value for `zulipVersion`