msglist: Remove buggy mute-state check in getFirstUnreadIdInNarrow helper.
This was causing an instance of #3457, "Narrow sometimes lands far
in the past".

`getFirstUnreadIdInNarrow` already only considers the "shown"
messages in a narrow, i.e., those returned by
`getShownMessagesForNarrow`. `getShownMessagesForNarrow`'s filtering
is entirely determined by a call to `shouldBeMuted`, passing the
narrow in question. It's a little hard to read, but from what I can
tell, the logic in that call is correct. (Even if it's not correct,
it's equivalent to the filtering used to populate the MessageList;
`getShownMessagesForNarrow` is also used for that.)

But then, `getFirstUnreadIdInNarrow` takes those correctly filtered
results, and, through the single-call-site functions
`findFirstUnread` and `isMessageRead`, makes an *additional* call to
`shouldBeMuted` — with the difference that HOME_NARROW is always
passed for `narrow`, instead of the narrow in question. It's like
this call is faking that we're in the HOME_NARROW, and that seems
unwise, here, a priori.

(The part where the `read` flag is considered is in
`isMessageRead`; that function returns `shouldBeMuted(message,
HOME_NARROW, subscriptions, mute) || !!flags.read[message.id]`.)

When the narrow in question is a muted stream, that `shouldBeMuted`
call returns true for all messages; this is consistent with the
desire not to see messages from a muted stream when you're
*actually* in the HOME_NARROW, but it isn't obviously necessary
here. That makes `isMessageRead` return true for all messages,
`findFirstUnread` returns undefined, `getFirstUnreadIdInNarrow`
returns null, so nextProps.initialScrollMessageId is null, and a
`scroll-to-anchor` event is sent to the WebView with null for
`scrollMessageId`.

It looks like the offending call to `shouldBeMuted` was introduced
in 0e7348986, but I see very little detail about the reasoning.

So, remove that call. If we're already displaying a message in the
message list (i.e., our filtering for display has already decided
it's not a muted message), and we point to it and ask, "Has the user
read this message?", we don't want the answer to be "Yes!" unless
it's actually marked as read.

Also, remove the tests of `isMessageRead` that correspond to the
undesired behavior.