eslint [nfc]: Indent block-style list items in ESLint config.
I found the "overrides" block at the bottom of .eslintrc.yaml a bit
confusing to see the structure of, so that I really wanted those two
items indented to show their relationship with the `overrides:`.

Then, I don't have a strong preference in other cases of this YAML
syntax (like for passing `error` and some options to a rule), but it
seems like we might as well pick a single indentation and stick with
it.  So, do that.  I find I mostly like this a bit better in the rules
cases too, especially on the complex config for no-restricted-syntax.

Use `git diff -b` or `git diff -w` to see this just indents and
adjusts a few related bits of formatting, and one comment.