presence: Simplify to just one way to track recent call.
We're using these two conditions to make essentially the same
decision: does this `start` call need to take care of causing
appropriate `doCallback` calls to happen, or has a previous call
(from before a previous `stop`) taken care of that?

The logic is easier to reason about if they make the decision in
the same way; so adjust one of them to work the same as the other.
This also lets us simplify the state by dropping `_previousTime`
entirely.

In most cases these conditions give the same answer, so that this
change is NFC.  In the cases where it does have an effect, the new
behavior seems more appropriate:

 * time hasn't passed, but _intervalID already null: This should
   be impossible.  The interval ID is cleared by `doCallback`, and
   `setInterval` shouldn't cause an early call, only a late one.

 * time has passed, but _intervalID not yet null: This can happen
   if the `doCallback` call from the interval is delayed.  In this
   case that `doCallback` call is imminent, so calling it again
   out of band may cause two calls in quick succession.

   If we're so starved for resources that the interval callback is
   going to be delayed much longer still... well, that's no good,
   but it's a situation affecting the app as a whole.  It wouldn't
   overall be helped by having individual bits of code try to
   notice and claim more resources for themselves.

Also edit the jsdoc corresponding to this logic, to be more
explicit about what happens in all cases.