deps: Add and link react-native-simple-toast@^0.1.1.
This package has support for CocoaPods at v0.1.1, with v1.0.0
requiring the use of CocoaPods; @remobile/react-native-toast
(currently used) does not have a .podspec file at all. Also,
@remobile/react-native-toast looks like it hasn't been touched in
three years. (There's even an old, open PR to add a .podspec, but it
hasn't been done.)

So, run `yarn add` and `react-native link` for
react-native-simple-toast@0.1.1. Also, translate the type annotation
from index.d.ts into a flow-typed stub generated with `flow-typed
create-stub`, and mock it in jest/jestSetup.js so tests don't fail
in the absence of NativeModules.LRDRCTSimpleToast.

Changing the call sites is left for the next commit in this series;
this commit only adds the dependency and makes it available, without
using it.

We won't take advantage of the CocoaPods support until #3983, "Set
up CocoaPods", is finished, but having all of our dependencies
support CocoaPods is crucial toward that work.

This package was in use before, but we replaced it with
@remobile/react-native-toast in f1356550b, and some remnants were
removed much later in 5e45ef2dd.

It's not very clear why the switch was made, in f1356550b, but the
description of #742 suggests that toasts weren't showing up on iOS;
that was with react-native-simple-toast@0.0.6 exactly (no caret in
package.json). There was a fork of this project, from
https://github.com/xgfe/react-native-simple-toast to
https://github.com/vonovak/react-native-simple-toast, and the latter
is used to build the NPM package with the same name we use in this
commit, which is also the name used in package.json before
f1356550b. An unmaintained repo (the repo from xgfe hasn't been
modified for about 4 years, according to GitHub) would have been a
strong motivator to switch to a different package, and it's possible
that the NPM package itself changed ownership, or it was mistakenly
assumed that the NPM package was linked to the xgfe repo instead of
the vonovak one.