iOS deps: Add expo-apple-authentication at latest.
Following their instructions for a "bare" app (as opposed to
"managed" by Expo) [1], which include the addition of the "Sign in
with Apple" entitlement, which we did through the Xcode UI as
instructed.

We don't have to worry about excluding this package from being
linked on Android, it seems; its `unimodule.json` specifies
`"platforms": ["ios"]`.

Empirically (when running Jest), there seems to be uncompiled
JavaScript in this package, so add an entry in
`transformModulesWhitelist`. Also mock the module, as we then get
errors about things being undefined.

[1]: https://github.com/expo/expo/tree/1c5eb9818/packages/expo-apple-authentication