StreamItem: Confirm '' cases are irrelevant, remove check, add JSDoc.
A response to

// TODO: confirm these '' cases are irrelevant, and remove.

, added in 84977a9cd.

With an effort, which was due to the somewhat convoluted plumbing, I
just verified that, when we *do* pass a value for the `color` or
`backgroundColor` prop to StreamItem, it is always the `color`
property from a Subscription. We'd like to add this to an explicit
contract for these props.

Greg reports [1] that the `!== ''` checks from 84977a9cd were there
because we weren't sure whether there was an intentional use of an
empty string in our code, to stand in for a missing/nullish color.
Now that we know that there isn't, we can remove the checks.

Contracts for `isMuted` and `isPrivate` are similarly induced from
the call sites.

So, add a JSDoc to that effect, to (paraphrasing Greg [1] again)
mark, for the sake of the implementation and the call sites, of what
they should rely on and what would be a bug.

As Greg points out, this isn't as good as making these requirements
inherent in the way the API works. We'd like to use a prop like
`subOrStream: Subscription | Stream` and glean data like `name` and
`isMuted` from that. Then the nested ternaries in `iconColor` and
`textColor` (which I'm still not sure I fully understand, even after
some pure refactors in recent commits) can be much simplified with,
e.g., a prop with something like an enum: `colorStyle: 'background'
| 'icon' | 'none'`.

But this better solution will be much more accessible after (or as
part of) #3767. Currently, among StreamItem's three call sites, its
data isn't only provided by a Stream or a Subscription object, but
also a "PseudoSubscription" (I'm not sure what that is) and an
UnreadStreamItem, so a relatively simple `subOrStream` prop isn't
feasible.

[1]: https://github.com/zulip/zulip-mobile/pull/4046#issuecomment-618044109