deps: Add local copy of redux-persist@4.10.2.
In #2762 (never merged), we tried to use v5 of redux-persist and
found that it had breaking changes that unacceptably reduced our
flexibility with migrations, so we decided never to use it, and
instead to vendor it and make our own changes as needed.

More background on the limitations of v5 is at
https://chat.zulip.org/#narrow/stream/243-mobile-team/subject/redux-persist.20v5/near/624465.
Its new way of doing migrations means you can't write a migration
that is aware of the entire state tree, unless you limit yourself to
storing the entire tree in a single key-value entry. We need to be
able to split it up for performance (avoiding a rewrite to disk of
everything each time a small change is made) and because each
key-value pair is has a hard limit of 2MB, after which the data is
just dropped. Even after compression, our entire state tree would be
within a factor of 4 of this limit, which is unacceptable.

More generally, the redux-persist maintainers seem not to have deep
familiarity with apps that depend on correct, effective migrations
to persist data across updates.

Our immediate need is to incorporate the changes from
rt2zz/redux-persist@a94f29139 and rt2zz/redux-persist@13a0b600b (not
released until 6.0.0) so that we can pass custom
serialize/deserialize functions for replace/revive logic so we can
store non-serializable data (such as a ZulipVersion) in Redux.

PR #2915 was opened a while ago to vendor redux-persist, but it grew
stale and gathered merge conflicts.

First, ensure that the version of redux-persist in node_modules is
4.10.2, which is what we'd been using. Then, copy the contents of
`node_modules/redux-persist/src` to `src/third/redux-persist` (this
contains all the needed source code) and copy
`node_modules/redux-persist/es/index.js.flow` there too; it has the
needed type definitions.

Then, address redux-persist's two dependencies that weren't listed
in our own package.json:

- `json-stringify-safe`: Add to package.json.

- `lodash`: The only bit of lodash that it uses is `isPlainObject`,
so add `lodash.isplainobject` to the package.json and make this
change, the only change to our vendored copy's source code:

  In `src/third/redux-persist/utils/isStatePlainEnough.js:

  ```diff
  - import isPlainObject from 'lodash/isPlainObject'
  + import isPlainObject from 'lodash.isplainobject'
  ```

Then run `yarn`.

Rather than fixing all the issues ESLint and Prettier flagged, tell
those tools to ignore them -- we didn't introduce them, and we'd
rather not accidentally introduce bugs by even these small changes.
Some of the output indicated potential code-quality issues (see note
at `no-use-before-define` in .eslintrc.yaml), some, just differing
style choices (e.g., no semicolons), and others, just the fact that
redux-persist wasn't using the very latest JavaScript features.