subscriptions reducer test: Test for `subscriptions` being empty array.
The functionality in the REALM_INIT case is straightforward, but an
empty array is a notable value, and we should pin down the behavior
for that value.

This is not quite identical to the test below it, titled 'when no
`subscriptions` data is given reset state', which doesn't do the
above: it tests what happens when `subscriptions` is falsy, not when
it's an empty array.

That test isn't really useful because `subscriptions` is never
missing. We remove that test and an unnecessary check in our live
code, in an upcoming commit, along with more rigorous argument for
the claim that it's dead code.