unread [nfc]: Revert unnecessary portions of last revert.
The commit 084d726fc "pms / recipient: Revert PR #4040."
reverted a series of 6 earlier commits:

  a111cfcf1 recipient [nfc]: Add minor documentation clarification.
  d8f45122e recipient [nfc]: Simplify pmUnreadsKeyFromMessage a bit.
  d464e33ec recipient [nfc]: Stop using email in pmUnreadsKeyFromMessage.
  79c867166 pm-conversations [nfc]: Get ownUser rather than just ownEmail.
  423cd19b1 unread [nfc]: Give getRecipientsIds a more informative name, and some jsdoc.
  2a2a2e130 unread [nfc]: Handle self-1:1 thread explicitly in getRecipientsIds.

But the bug being addressed was in specifically commit 79c867166,
so there's no need to revert the two commits that preceded it.
There's also no need to revert a111cfcf1, a documentation fix that
doesn't interact with the code added in 79c867166.  So, restore
those changes.