realm reducer [nfc]: Completely replace state on REALM_INIT.
The point of this state subtree is to carry miscellaneous data from
the server; it doesn't retain any state across a `/register` call,
and if it did that would probably be a bug.

Drop the spread of the old state, to make that explicit and to get
the type-checker to check it for us.