register [nfc]: Inline effect of calling `objectToParams`.
This is the last remaining call site of `objectToParams`. Inline its
effect, to enable its removal in an upcoming commit.

(Note that it's possible that some of the now-explicitly-extracted
values are not present in the `params` object, so there may be a
subtle difference in the resulting argument to `apiPost`: some
formerly absent keys will be passed in as present with the value
`undefined`. This is not a functional difference in this context:
`apiPost` and its siblings filter that out.)