android build [nfc]: Don't take `debug.keystore` config in RN v60 upgrade.
This commit marks part of the RN v0.59 -> v0.60 changes to the
template app [1] that we don't take. This commit may freely be
positioned before the upgrade.

Corresponds to facebook/react-native@63ebbd6bf, released in RN
v0.60.0.

Android Studio already takes care of supplying a debug keystore
automatically, and we haven't seen any issues with that, nor do such
issues seem to be a motivation for that upstream change.

Probably, they wanted to make it easier to configure *release*
keystores. But we already have a process for that [2], which works
better for us than anything like what's proposed here.

This change was also pretty turbulent, as it had to be partly
amended in facebook/react-native@d55025694b [3]. That fix wasn't
released until a full version later, in RN v0.61. Hundreds of people
who had followed the RN v60 upgrade guide commented and reacted on
the issue, having had to manually add to their `.gitignore` and
download and copy a `debug.keystore` file.

So, don't take this change. See also discussion [4].

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.59.10&to=0.60.6

[2]: https://github.com/zulip/zulip-mobile/blob/master/docs/howto/release.md#prepare-android

[3]: https://github.com/facebook/react-native/issues/25629

[4]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Android.20signing.20%28RN.2060.20upgrade%29/near/880752