message actions: Use authless URLs to access uploaded files.
Using the new helper tryGetFileDownloadURL, we now try to generate
and use temporary authless URLs for realm uploaded files.

We open files using the temporary URL in the browser, as opposed
to downloading them directly, because 1. The browser can usually
open several file types instantly and 2. If the browser can not handle
displaying the file, a system dialog shows up, both on Android and iOS,
showing platform specific ways to handle the file, such as opening in
Google Drive, Saving to iCloud and other options based on installed
apps, which is convenient for the user.

[greg: On iOS this doesn't actually work yet, but the bug it runs into
 is unrelated to #3303.]

Fixes: #3303