redux-persist types: Remove deprecated $Subtype utility.
Preparatory to the RN v0.59 -> v0.60 upgrade [1].

This will be flagged with a `deprecated-utility` error when we
upgrade to Flow 0.98.0. [2]

I don't see why this was `$Subtype<string>` instead of just `string`
in the first place, anyway. It's not clear to me what work is
supposed to be done with `$Subtype`. The docs [3] don't define a
limit to the strings one can use as a consumer of this API. This is
our use of `whitelist`:

```
const reduxPersistConfig: Config = {
  whitelist: [...storeKeys, ...cacheKeys],
  // etc.
};
```

where `storeKeys` and `cacheKeys` are both typed as
`Array<$Keys<GlobalState>>`.

I suppose an element of the array we pass as `whitelist` is indeed a
subtype of `string`, being limited to particular strings (the keys
of GlobalState), but it's not clear that this *must* be the case,
from `redux-persist`'s perspective.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.59.10&to=0.60.6

[2]: https://flow.org/en/docs/linting/rule-reference/#toc-deprecated-utility

[3]: https://github.com/rt2zz/redux-persist#blacklist--whitelist