lint: Prepare Airbnb overrides/adjustments for upcoming ESLint upgrade.
Some more notes about the current state of two of these:

- react/static-property-placement
- react/state-in-constructor

If you look in node_modules/eslint-config-airbnb/rules/react.js,
you'll see TODOs that would bring Airbnb's rules in line with what
we've chosen, as soon as the relevant language features are
available in `babel-preset-airbnb`. Those features are available for
us already, so take the desired changes.

See discussion on all of these rules at and after
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4120.20ESLint.206.2E8.2E0.20upgrade/near/888174.