docs/build-run: Drop the Windows Command Prompt; prefer Git Bash.
In principle it shouldn't be hard to support at least running a Gradle
build (and so `react-native run-android`) from CMD.EXE aka the Windows
Command Prompt -- the actual scripts can use Git Bash, and we just need
to teach our build-webview Gradle task to find it.

But in practice, it's difficult to keep it working because none of our
regular contributors for the mobile app use Windows.  Keeping the
build working from Git Bash is a lot more tractable because it's much
more similar to the environments our core team does use (GNU/Linux and
macOS) and so just doesn't tend to break.  And, on the other hand, Git
Bash is right there for people to use -- installing it is already in
our setup instructions for Windows -- so nobody's left out by not
supporting something else.