deps: Remove direct `eslint-config-airbnb-base` dependency.
We don't ever import this ourselves, and `eslint-config-airbnb`
already has it, since it's listed as a `dependency` by that package.

It's never been listed there as a peer dependency, even as far back
as v15.0.2 [1], the version we were on when we added
`eslint-config-airbnb-base` as a direct dependency, in 984f77157. A
listing in `peerDependencies` would be a compelling reason for us to
depend on it directly [2].

Moreover, 984f77157 didn't explain the addition, and the main focus
of that commit was elsewhere, so it seems unlikely that it was
well-considered.

So, remove it. This will simplify ESLint upgrades, like the one
later in this series.

[1]: https://github.com/airbnb/javascript/blob/c68fa632624f088b4bbe81192a50554b0c6af970/packages/eslint-config-airbnb/package.json

[2]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/unimodules.20unmet.20peer.20dependency/near/850108